Code cleanup.  Addresses https://github.com/CompuCell3D/CompuCell3D/issues/459
