+++ cc3d/player5/UI/CellTypeColorMapView.py
@@ -1,10 +1,9 @@
-
 from PyQt5.QtCore import *
 from PyQt5.QtGui import *
 from PyQt5.QtWidgets import *
 import sys
-from .checkbox_delegate import CheckBoxDelegate, ColorEditorDelegate
-from .qcolor_delegate import ColorDelegate
+from .checkbox_delegate import CheckBoxDelegate
+from .color_delegate import ColorDelegate
 
 
 class CellTypeColorMapView(QTableView):
@@ -13,27 +12,19 @@ class CellTypeColorMapView(QTableView):
         QTableView.__init__(self, parent)
         self.setFrameStyle(QFrame.NoFrame)
         self.setSelectionBehavior(QAbstractItemView.SelectRows)
-        # self.setColumnWidth(0, 30)
-        # self.setColumnWidth(1, 50)
         self.setAlternatingRowColors(True)
         self.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
         self.horizontalHeader().setStretchLastSection(True)
         self.setEditTriggers(QAbstractItemView.NoEditTriggers)
         self.setSelectionMode(QAbstractItemView.NoSelection)
 
+        # note - delegates need to be member of a class to avoid them being garbage-collected. This would cause
+        # cryptic crashes
         self.check_box_delegate = CheckBoxDelegate(None)
         self.setItemDelegateForColumn(3, self.check_box_delegate)
 
-        # color_delegate = ColorDelegate()
-
-        self.color_delegate = ColorEditorDelegate(None)
+        self.color_delegate = ColorDelegate(None)
         self.setItemDelegateForColumn(1, self.color_delegate)
-
-
-        # self.setColumnWidth(0, 30)
-        # self.setColumnWidth(1, 50)
-
-
 
         # on OSX we do not resize row height, we do it only on windows and linux
         if not sys.platform.startswith('darwin'):
@@ -42,9 +33,8 @@ class CellTypeColorMapView(QTableView):
             # verticalHeader.setResizeMode(QHeaderView.Fixed)
             # verticalHeader.setDefaultSectionSize(20)
             
-        # vm - viewmanager, instance of class TabView
+        # vm - viewmanager, instance of class SimpleTabView
         self.vm = vm
-        #self.__resizeColumns()
 
     def update_content(self):
 
@@ -56,25 +46,13 @@ class CellTypeColorMapView(QTableView):
         model.read_cell_type_color_data()
         model.endResetModel()
 
-
     def mousePressEvent(self, event):
-
-        # pg = CompuCellSetup.persistent_globals
-        # if pg.steering_panel_synchronizer.locked():
-        #     return
 
         if event.button() == Qt.LeftButton:
             index = self.indexAt(event.pos())
             model = index.model()
             if index.column() == model.color_idx:
                 self.edit(index)
-            # if index.column() == model.show_in_3d_idx:
-            #     print('trying to edit show in 3d')
-            #     self.check_box_delegate.editorEvent(event=event, model=model, option=None, index=index)
-            #     # event.ignore()
-            #     # self.edit(index)
-
-
         else:
             super(CellTypeColorMapView, self).mousePressEvent(event)
 
@@ -86,28 +64,7 @@ class CellTypeColorMapView(QTableView):
             if index.column() == model.show_in_3d_idx:
                 # somewhat hacky solution - editor event checks also for mouseRelease event so we need to
                 # reimplement mouseReleaseEvent
-                self.check_box_delegate.editorEvent(event=event, model=model, option=None, index=index)
+                option = self.viewOptions()
+                self.check_box_delegate.editorEvent(event=event, model=model, option=option, index=index)
         else:
             super(CellTypeColorMapView, self).mouseReleaseEvent(event)
-
-
-    # def setParams(self):
-    #     """
-    #     Sets the parameters if the QTableView when the model is set
-    #     """
-    #     if self.model() is None:
-    #         return
-    #
-    #     # assert self.model().
-    #     # print self.model()
-    #     # import sys
-    #     # if not sys.platform.startswith('darwin'): # on OSX we do not resize row height
-    #         # for i in range(0, self.model().rowCount()):
-    #             # self.setRowHeight(i, 20)
-    #
-    #     self.setColumnWidth(0, 130)
-    #     #self.cplugins.setColumnWidth(1, 200)
-    #     self.setAlternatingRowColors(True)
-    #     self.horizontalHeader().setStretchLastSection(True)
-
-
